+++ check/contents/check_timeouts_section.go
@@ -1,0 +1,11 @@
+package contents
+
+func (d *Document) checkTimeoutsSection() error {
+	section := d.Sections.Timeouts
+
+	if section == nil {
+		return nil
+	}
+
+	return nil
+}